singleページのレイアウト修正

+++ app/View/Post/index.ctp
@@ -3,7 +3,7 @@
 <h2>Eveetする！</h2>
 <?php
     echo $this->Form->create('Posts', array('action' => 'add_post'));
-    // echo $this->Form->hidden('user_id', array('value' => $user['id']));
+    echo $this->Form->hidden('user_id', array('value' => $user['id']));
     echo $this->Form->textarea(__('content'), array('rows'=> 7,'class' => '', 'placeholder' => 'いまどうしてる？'));
     echo $this->Form->submit(__('Submit'), array('class' => ''));
     echo $this->Form->end();
@@ -15,7 +15,8 @@
 	echo $post['Post']['content'];	
 	echo $post['Post']['created'];
 	echo $post['User']['username'];
-	echo $post['Post']['delete_flag'];
-	echo "<a href='/evis/posts/delete_post?post_id=" . $post['Post']['id'] . "'>削除</a><br />";
+	if($user['id'] == $post['Post']['user_id']){
+		echo "<a href='/evis/posts/delete_post?post_id=" . $post['Post']['id'] . "'>削除</a><br />";
+	}
 	endforeach;
 ?>
